+++ ui/components/multichain-accounts/multichain-account-list-menu/multichain-account-list-menu.tsx
@@ -32,7 +32,7 @@ import {
   ACCOUNT_OVERVIEW_TAB_KEY_TO_TRACE_NAME_MAP,
   AccountOverviewTabKey,
 } from '../../../../shared/constants/app-state';
-import { getWalletsWithAccounts } from '../../../selectors/multichain-accounts/multichain-accounts-selectors';
+import { getWalletsWithAccounts } from '../../../selectors/multichain-accounts/account-tree';
 import { MultichainAccountsTree } from '../multichain-accounts-tree';
 import { AccountMenu } from '../../multichain/account-menu';
